+++ zarr/__init__.py
@@ -10,6 +10,7 @@ from zarr.creation import (array, create, empty, empty_like, full, full_like,
 from zarr.errors import CopyError, MetadataError
 from zarr.hierarchy import Group, group, open_group
 from zarr.n5 import N5Store, N5FSStore
+from zarr._storage.store import v3_api_available
 from zarr.storage import (ABSStore, DBMStore, DictStore, DirectoryStore,
                           KVStore, LMDBStore, LRUStoreCache, MemoryStore, MongoDBStore,
                           NestedDirectoryStore, RedisStore, SQLiteStore,
@@ -19,3 +20,8 @@ from zarr.version import version as __version__
 
 # in case setuptools scm screw up and find version to be 0.0.0
 assert not __version__.startswith("0.0.0")
+
+if v3_api_available:
+    from zarr._storage.v3 import (ABSStoreV3, DBMStoreV3, KVStoreV3, DirectoryStoreV3,
+                                  LMDBStoreV3, LRUStoreCacheV3, MemoryStoreV3, MongoDBStoreV3,
+                                  RedisStoreV3, SQLiteStoreV3, ZipStoreV3)
