+++ zarr/tests/test_attrs.py
@@ -4,7 +4,8 @@ import pytest
 
 from zarr._storage.store import meta_root
 from zarr.attrs import Attributes
-from zarr.storage import KVStore, KVStoreV3
+from zarr.storage import KVStore
+from zarr._storage.v3 import KVStoreV3
 from zarr.tests.util import CountingDict, CountingDictV3
